MTDSA-5664 Added Response Model for Amend Insurance Policies + Updated file structure
