+++ app/v1/connectors/RetrieveSavingsConnector.scala
@@ -21,7 +21,7 @@ import javax.inject.{Inject, Singleton}
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.play.bootstrap.http.HttpClient
 import v1.models.request.savings.DeleteRetrieveRequest
-import v1.models.response.retrieveSavings.RetrieveSavingsResponse
+import v1.models.response.savings.retrieveSavings.RetrieveSavingsResponse
 
 import scala.concurrent.{ExecutionContext, Future}
