+++ app/v1/models/response/savings/retrieveSavings/ForeignInterest.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package v1.models.response.retrieveSavings
+package v1.models.response.savings.retrieveSavings
 
 import play.api.libs.functional.syntax._
 import play.api.libs.json.{JsPath, Json, OWrites, Reads}
