+++ app/v1/services/AmendSavingsService.scala
@@ -26,7 +26,7 @@ import v1.controllers.EndpointLogContext
 import v1.models.errors._
 import v1.models.outcomes.ResponseWrapper
 import v1.models.request.savings.amend.AmendSavingsRequest
-import v1.models.response.amendSavings.AmendSavingsResponse
+import v1.models.response.savings.amendSavings.AmendSavingsResponse
 import v1.support.DesResponseMappingSupport
 
 import scala.concurrent.{ExecutionContext, Future}
