+++ test/v1/models/retrieveSavings/RetrieveSavingsResponseSpec.scala
@@ -19,6 +19,7 @@ package v1.models.response.retrieveSavings
 import play.api.libs.json.{JsError, JsValue, Json}
 import support.UnitSpec
 import v1.fixtures.RetrieveSavingsFixture._
+import v1.models.response.savings.retrieveSavings.RetrieveSavingsResponse
 
 class RetrieveSavingsResponseSpec extends UnitSpec {
